Add new GCC built-in definitions
I tried to generate the call graph for the last version of Coreutils
using csmake to generate the processing script automatically. But when I
executed `cscout -R cgraph.txt?all=1 make.cs` to generate the call graph
it failed with:

```
/usr/include/stdlib.h:140: error: syntax error
The text leading to this error is: [extern _Float32 strtof32]
```

Adding `#define _Float32 float` to include/template/gcc-defs.h solves this
problem.

The definitions for GCC can be found in chapter 5-6-7 in the following
link.

<https://gcc.gnu.org/onlinedocs/gcc/>