Add hard_defined pragma to handle special cases
When tried to generate the call graph for the last version of Coreutils,
using csmake to generate the processing script automatically, the
following error occurred.

```
/usr/include/stdlib.h:76: error: syntax error
The text leading to this error is: [__extension__ typedef]
```

To solve this error normally, we'll add the `#define __extension__` to
`include/template/gcc-defs`, but when we did it, the error didn't
resolve. The reason was that somewhere in Coreutils, there was an
`undef` for `__extension__`. To handle that, we need to block `undef`
for some specific defines. To do so, we added a new pragma directive,
called hard_defined.