Fix bug and add one more option to -R option
When we called cscout with the `-R cgraph.txt?n=p` the `n=p`
does not have any impact because in the `function_label` function the
cgraph_show is read from `Option::show_function_type`. To fix this issue
I set the value of `n` with a new setter function directly on
`Option::cgraph_show`.

The new option is the `show_function_type` option with which a prefic of
`static:` or `public:`, based on function type,  is added to nodes in
call graphs. To enable it from the command line you should pass `type=1`.
For example, `cscout -R cgraph.txt?n=p&type=1 make.cs`