Add new options.
The functionality of option -T is now separated in -T itself, -s and -k.
The first one is used to set and use a given deirectory as the temporary
directory, the second one is used to generate a .cs file per project and
the third one to keep the temporary directory in place.
Moreover, -d is added to not only keep the temporary directory but also
to print more debug messages.