Fix wrong database identification in tokens
Rather than identify the database to merge tokens (which is difficult,
because there is no direct map), do away with numbered databases for the
eid_to_global_map and have only the current database and null for the
older one.  This requires renumbering all eids to avoid clashes.

Adjust other dependent files, and follow the same principle for
function-ids.