Fix handling of function ids
The previous method based on same functionid sequences was incorrect,
because their ECs could be merged together for different functions, as
was the case, for example, for __io_bw() in Linux's
arch/arm64/include/asm/io.h and arch/riscv/include/asm/mmio.h.

Therefore, revert to the use of a heuristic based on function location
and name for allocating and mapping function ids.
This removes from CScout the responsibility of mapping function ids.