Merge functionds based on Tokids not ECs
Tokids are used internally for comparing functions in the absence of a
symbol table, whereas ECs can bring confusion because the bring
semantically unrelated tokids together.

It turns out that the Functionid table also needs to include the length
of each part, so add this to it and reseed the test data.