Use recursive invocations for macro expansion
This was done in an attempt to fix macro caller in macro expansion.
Without it, expanding a macro with two macros in it, wrongly identifier
the first as the called to the second.

The unsuccessful fix required refactoring the macro_expand function to use
Prossner's recursive implementation with the caller passed and adjusted
as part of the recursive invocation.